refactor: moved helpers to app folder
